+++ src/Boot/src/BootloadManager/StrategyBasedBootloadManager.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Spiral\Boot\BootloadManager;
 
+use Spiral\Boot\Attribute\BootloadConfig;
 use Spiral\Core\ScopeInterface;
 
 final class StrategyBasedBootloadManager extends AbstractBootloadManager
@@ -19,7 +20,7 @@ final class StrategyBasedBootloadManager extends AbstractBootloadManager
     /**
      * Bootload all given bootloaders.
      *
-     * @param array<class-string>|array<class-string, array<string,mixed>> $classes
+     * @param array<class-string>|array<class-string, array<string,mixed>|BootloadConfig> $classes
      *
      * @throws \Throwable
      */