+++ src/Boot/src/BootloadManagerInterface.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Spiral\Boot;
 
+use Spiral\Boot\Attribute\BootloadConfig;
 use Spiral\Boot\Bootloader\BootloaderInterface;
 
 /**
@@ -27,7 +28,7 @@ interface BootloadManagerInterface
      *    CustomizedBootloader::class => ["option" => "value"]
      * ]
      *
-     * @param TClass[]|array<TClass, array<string,mixed>> $classes
+     * @param TClass[]|array<TClass, array<string,mixed>|BootloadConfig> $classes
      * @param array<\Closure> $bootingCallbacks
      * @param array<\Closure> $bootedCallbacks
      *